requirements.txt
requests==2.31.0
python-dotenv==1.0.1
pytz==2025.1
ollama==0.1.6
elevenlabs==0.2.27
pygame==2.5.2